Fix case insensitivity issues (#563)
Without this change it's impossible to reference two Postgres tables in
the same query that only differ in casing. This fixes that by not using
a `case_insentive_map` to track the Postgres table and schema names.
While this doesn't completely fix all issues caused by case insentivity
mismatch between Postgres and DuckDB it does fix the major one described
in #116.

Fixes #116