Add tests for assigning aliases of different casing
With the `read_parquet` syntax from before #531 we had a bug where it
was impossible to change casing of an alias from upper to lower:

```
select upper as "upper", lower as "LOWER" from read_parquet('/tmp/tmp.parquet') as t(upper int8, lower int8) limit 2;

 UPPER | LOWER
-------+-------
     0 |     0
     1 |     1
(2 rows)
```

See #564 for more details.

Luckily #531 has completely fixed this problem, so all this PR does is add
some regression tests.

Finally it also corrects the output file of the `case_insensitivity.sql`
test, which was introduced by #563. It also adds that test to the
`schedule` file, which was the reason that CI didn't notice that the
output was wrong.

Fixes #564
Related #563