Build images on dedicated runners (#578)
Docker build was recently failing because building an image for the
ARM64 CPU on an AMD64 runner was prohibitively slow.

In this PR, we build each image on a dedicated runner matching the
target platform.

In a following one, we will merge the two images and push them under an
unified tag.