Don't sync MotherDuck metadata forever (#582)
Keeping a connection open to MotherDuck forever is undesirable from a
resourcing perspective. This changes that to behave similarly to the
DuckDB CLI: After 5 minutes (by default) syncing will stop. The main
difficulty is to restart the syncing. In the DuckDB CLI the syncing will
start again after some activity is detected, but the background worker
never triggers activity. So instead we need to make sure that activity
in other connections triggers a restart of the syncing in the background
worker. This is done using some very simple IPC in shared memory.