Fix interruption crash (#584)
When the "Query cancelled" exception is thrown (following a SIGINT
signal), various destructors are called, among which
`PostgresTableReader`'s which cleanup the PG state. If an exception is
thrown during the stack unwinding and call to PG function, it results in
an undefined behavior which materialize as a process crash.

So to avoid that, we eagerly consume the pending tasks before throwing
the exception.