Print function name rather than caller (#586)
When an error is raised, we used to print the `__func__` string, which
contains the current function name where the macro is placed, hence the
actual "caller" of the function being wrapped.

This PR replaces the caller name with the name of the function being
wrapped.

In an initial version I was displaying both, but found it to be
redundant.

An alternative would be to check if caller name contains the function
name (or vice-versa) which happens often when we have functions such as
`TopLevelFunc` and `TopLevelFunc_cpp`, and:
* if true, only display one
* if false, display both

But that sounds a bit much.

Happy to hear some feedback :-)