Add ability to link libduckdb statically
pg_duckdb is not the only Postgres extension using DuckDB these days. If
two extensions both dynamically link to libduckdb.so and install that
library into the Postgres libdir, then it is a race to see which
extension installs their libduckdb last. This can lead to undefined
symbols in any of the DuckDB Postgres extensions.

Signed-off-by: Tristan Partin <tristan@partin.io>