Handle PG columns with arrays with 0 dimensions correctly (#616)
As was found in #556 sometimes array columns in Postgres are listed to
have 0 dimensions, that's obviously because of missing information. This
PR starts assuming that those arrays have a single dimension (because
single dimension arrays are the most common type).

Given that that guess can obviously be wrong too. This adds the
limitations that our multi-dimensional array support currently has in
our docs.

Fixes #556