Fix crash when enabling MotherDuck (#631)
With #630 I accidentally broke MotherDuck support. I had only tested
that it would not crash with an incorrect MotherDuck token, but I had
not confirmed that a correct one would continue to work correctly. This
fixes that.

The reason that this broke was that we would hold a raw pointer to a
unique_ptr that was already out of scope. This fixes that by creating
the unique_ptr in the correct scope again.