Add additional functionality for the duckdb.unresolved_type
In 0.3.0 we added the `duckdb.unresolved_type` type and implemented a
bunch of operators, aggregates and functions for it. Ofcourse the list
will always be incomplete and casts to actual types might always be
necessary in some cases. But for usability we should still try to
implement the most commonly used functionality. So this adds a bunch of
additional things that I ran into over the past weeks:

1. The `min` aggregate, this was not present in 0.3.0 by accident.
   Probably because duckdb has two `min` aggregates.
2. Casts to `time`, `timetz`, `interval` and `bit`
3. Support for `date_trunc`, `length` and `regexp_replace`
4. Add general support for the `strftime` and `strptime` functions. This
   is useful even for already resolved/casted `timestamp`/`date` values
   because Postgres its date formatting functions like `to_char` are not
   available in DuckDB.