Remove code related to http caching
We had implemented our own http caching in #88 when nothing like that
existed yet. Now there are two different threads that are implementing
something like this in a better way than we have done. So this PR
removes our custom caching code from our codebase. Instead in a future
release we'll make sure that pg_duckdb works well with these at least
one of these other caching solutions.

For reference the two projects that are currently implementing caching
are:
1. An official in-memory cache by DuckDB, which will eventually improved
   to also support on-disk caching: https://github.com/duckdb/duckdb/pull/16463
2. A community provided extension called `cache_httpfs`: https://github.com/dentiny/duck-read-cache-fs