Start supporting DuckDB DDL in transactions (#632)
This starts supporting DuckDB DDL in transactions. There are still some
limitations, but the most common use cases should now work as expected.
For the cases that are disallowed, it's possible to disable the
mixed-writes detection by setting
`duckdb.unsafe_allow_mixed_transactions` to `true`. This is not
recommended for most usage, but it is a useful escape hatch for cases
where people cannot manually control the transaction end/start.

A major change this does in the detection is that it starts to
completely rely on CommandId for detecting Postgres writes. Previously
we also looked at WAL, but it turns out that is very unreliable, because
even read queries can trigger WAL writes (e.g. to update hint bits).