Add automated tests for MotherDuck integration (#649)
This adds some basic infrastructure to be able to test the MotherDuck
integration of pg_duckdb. It also includes a few very basic tests. For
now these tests are not run in CI yet, we can do that in a follow-up PR.
But at least you can manually run these tests from your own machine now.