Add additional functionality for the duckdb.unresolved_type (#643)
In 0.3.0 we added the `duckdb.unresolved_type` type and implemented a
bunch of operators, aggregates and functions for it. Of course the list
will always be incomplete and casts to actual types might always be
necessary in some cases. But for usability we should still try to
implement the most commonly used functionality. So this adds a bunch of
additional things that I ran into over the past weeks:

1. The `min` aggregate, this was not present in 0.3.0 by accident.
Probably because DuckDB has two `min` aggregates.
2. Casts to `time`, `timetz`, `interval` and `bit`
3. Support for `date_trunc`, `length` and `regexp_replace`
4. Add general support for the `strftime` and `strptime` functions. This
is useful even for already resolved/casted `timestamp`/`date` values
because Postgres its date formatting functions like `to_char` are not
available in DuckDB.
5. Operators needed to implement LIKE/ILIKE/SIMILAR TO, i.e. `~`, `~~`,
`~~*`, `!~`, `!~~`, and !~~*`
6. Add general support for the `epoch`/`epoch_ms`/`epoch_us`/`epoch_ns`
functions. These are useful for lots of data processing tasks, because
often timestamps are stored as a number relative to the epoch in data
files/tables.

To make the string of parsing/formatting of `strftime`/`strptime`
consistent this also starts syncing the PG timezone to DuckDB.