Start one background worker on first DuckDB query (#544)
Up until now, we were starting the background worker as part of the
extension.

The problem with this approach is that it imposes a unique worker, and
thus cannot updated multiple database.

In preparation to have one BGW per database, this PR triggers the start
of a background worker when `IsExtensionRegistered` is called and the
cache is populated.

We've prevented potential race conditions that would cause two bgw to
start by checking a lock on a temporary file.

Note: a [following PR](https://github.com/duckdb/pg_duckdb/pull/545)
will actually start one BGW per database

---------

Co-authored-by: Jelte Fennema-Nio <github-tech@jeltef.nl>