Fix server crash when EXPLAINing a prepared statement
With a specific query order it was possible that Postgres would cache
the plan for the EXPLAINed version of a query instead of the plan for
the actual query. This fixes that by only ever adding EXPLAIN to a query
when actually executing the query, not also when only planning it. This
way the plan that will be cached is always the actual query plan, but
at execution time that might be changed in case the query is EXPLAINed.