Fix server crash when EXPLAINing a prepared statement (#660)
With a specific query order it was possible that Postgres would cache
the plan for the EXPLAINed version of a query instead of the plan for
the actual query. This fixes that by only ever adding EXPLAIN to a query
when actually executing the query, not also when only planning it. This
way the plan that will be cached is always the actual query plan, but at
execution time that might be changed in case the query is EXPLAINed.

This also fixes two bugs involving EXPLAIN ANALYZE:
1. EXPLAIN ANALYZE + CTAS into a Postgres table is now disallowed when
using DuckDB execution. Before it was putting the plan into the table...
Not the result of the query.
2. Don't execute the EXPLAIN ANALYZE twice in DuckDB. For read-only
queries this would result in the query taking twice as long, while only
reporting the runtime for one of the queries. When using EXPLAIN ANALYZE
on DML statements this would actually matter a lot, because those would
have the side-effect twice, i.e. data would be inserted twice.

Fixes #410