Fix issue with too long schema names breaking MotherDuck sync
If you have a database+schema that would result in a `ddb$` schema with
more than 63 characters it would completely break the MotherDuck
background syncing, due to an error-restart-loop. This changes the code
to skip that schema.

It also codes a bit more defensively in the place where it was also
previously throwing the error continuously. If the schema cannot be found
when we try to add a dependency for it, we say that we failed to create
the schema and skip it instead of getting into this error loop.