Fix type mismatch issues with CTAS and materialized views (#706)
DuckDB and Postgres can return different types for the same queries.
This can confuse Postgres in various ways when these queries are
involved in CTAS or materialized views. This fixes a bunch of issues
related to that and adds appropriate regression tests.

This includes a fix for the materialized view issue that was found in
#583