Improve MD test user creation (#707)
In order to correctly test https://github.com/duckdb/pg_duckdb/pull/545
we will need to have multiple test users.

This PR paves the way to generate as many MotherDuck test user as we
want.

I am considering removing the support for `MOTHERDUCK_TEST_TOKEN`, but I
have kept it for now.

The main change is now we can run the test suite with a token that can
generate test users and it will generates one and use it through the
test suite.