Support subscripting on duckdb.map (#773)
In #689 the subscript operator was not implemented for the duckdb.map
type. This worked fine for MAPs returned by `duckdb.query()` because
that returned duckdb.unresolved_type. But not when getting a map from a
`duckdb` table column, i.e. a MotherDuck table. So this implements the
subscript operator directly on the `duckdb.map` type.

Fixes #732