Compatibility with timescale time_bucket function
The timescale extension includes a time_bucket function that conflicts
with our version. This changes our SQL creation scripts to ignore
creating the time_bucket functions in the public schema, if we detect a
conflict. It seems better to allow installing the pg_duckdb extension
partially like this than failing the installation completely.