Harden our CustomScan against unexpected type changes (#759)
This potential issue has been bothering me for a while now. It's
possible for DuckDB to change its mind about the columns that it returns
between the Postgres planning and execution phase. This change makes
sure we detect such cases with a clear error, instead of having Postgres
return weird results or crash.