Disallow DuckDB execution inside functions (#764)
pg_duckdb changes result types of queries at the planning stage. This is
needed for pg_duckdb to function correctly. However, Postgres code is
sometimes not written with this in mind. One such case is sql functions,
as reported in #749. There it causes a crash, because Postgres expects a
different type to be returned by the function than what is actually
returned.

As far as I can tell there is no easy way of fixing this, apart from
changing Postgres to learn how to deal with this. I'll open issue/patch
to Postgres for this at some point in the near future.

For now, we simply start disallowing all DuckDB execution inside
functions. That is quite a shame because using DuckDB execution in
functions works fine in many other cases. Sadly, I cannot think of a way
to only disallow the crashing cases. Since a very important part of
pg_duckdb 1.0 is stability, it seems better to err on the side of
caution and disable this feature for now and stabilize it in a future
release.

Fixes #749