tests: Handle DDL command in cur.sql() differently
For ease of programming we have a `cur.sql()` method that can be used to
execute any sql command. Fetching rows for DDL statements doesn't work
though, so we were catching that error and ignoring it. In psycopg 3.2.8
the exact error message was changed.

Instead of detecting the new error message, this instead makes sure that
we don't get this error at all. This is done by detecting whether our
result has rows, before calling `fetchall()`.