Fix `loop-until` (#790)
In a recent refactoring we changed the `loop_until` implementation using
a generator with `while`. This, in effect defeat the purpose of the
function since it will loop indefinitely, the generator being always
truthy.

This PR replace `loop_until` with a `wait_until` function everywhere
which checks if the given function becomes truthy eventually.