Don't convert unsupported numerics to double by default (#795)
This stops converting unsupported NUMERIC sizes to doubles. This
behaviour is still supported, but it is now opt-in. This means pg_duckdb
won't silently reduce precision. It also starts requiring users to
specify precision when using NUMERIC types in a table definition,
because the default precision is different between PostgreSQL and
DuckDB.

Closes #471

---------

Co-authored-by: dentiny <dentinyhao@gmail.com>