Fix CTAS with NUMERIC type
In #795 errors were added when using NUMERIC without a precision. It
turns out that these were impossible to avoid for a CTAS command,
because we didn't use the precision provided by DuckDB when creating the
Postgres table. This fixes that and adds some tests to encode the
current behaviour.

Related to #795