Use DROP IF EXISTS for when dropping DuckDB tables
It's possible for DuckDB and Postgres to get out of sync. Obviously we
don't generally want this, but in some cases it's impossible to avoid.
In those cases users should at least be able to clean up out-of-sync
tables themselves. This change allows users to drop the Postgres side of
tables, even if the DuckDB table does not exist.