Use Postgres builtin filesystem roles
This changes slightly how we determine if the local file system should
be accessible by DuckDB. Instead of defaulting the disabled_filesystems
GUC to `LocalFileSystem` we now default it to the empty string. To make
such a configuration secure, we now automatically add `LocalFileSystem`
to the configured list before configuring it in Postgres if the user
isn't a member of the `pg_read_server_files` and `pg_write_server_files`
roles.