Fix pgspot errors for 1.0.0 release (#817)
We weren't fully qualifying all functions and operators in our UTF8
encoding check. This sets a safe search_path before doing the check, so
that we don't accidentally use a function or operator from a different
schema during extension installation/upgrade.

Failing tests (after changing the search_path) notified me of the fact
that we also don't store some of our internal functions in the duckdb
schema. That was unintended. This fixes that too.