Add make_timestamp and make_timestamptz (#838)
On the DuckDB main branch `epoch_ms` is removed, and instead
`make_timestamp[tz]` is recommended:
https://github.com/duckdb/duckdb/pull/17816

So, this starts supporting make_timestamp[tz] in pg_duckdb. We might
even want to remove `epoch_ms` support, since it has not been released
yet, and DuckDB intends to remove it in a future release.