Remove bigint epoch_ms function
DuckDB 1.4 removes the bigint variant of epoch_ms for consistency:
https://github.com/duckdb/duckdb/pull/17816

This also removes that epoch_ms function for pg_duckdb, to avoid having
to do a breaking change in the next release.

This also removes the unresolved_type version of epoch_ms, as that can
be used inadvertently to call the bigint one. Once pg_duckdb starts
using DuckDB 1.4, we can add this back.