Don't call into the Postgres planner anymore when we take over
In the past we needed to call into the Postgres planner for permission
checking on Postgres tables. This has not been needed for a while now.
We were still doing it out of caution for breaking other things, but it
seems like that might be causing problems by itself too. And also it's
quite a waste of CPU cycles.