Wait for motherduck sync in docker image
After the Docker init.d scripts have run Postgres is restarted. Normally
that's would be fine, but when enabling MotherDuck we still need the
initial table sync to finish, otherwise the first queries that a user
tries will fail. This adds some logic to the motherduck init script to
wait until the sync has finished.