Detect unsupported Postgres types in more cases
Previously we would throw an error when a query returned an unsupported
Postgres type (for instance a too wide decimal). Turns out this was not
enough, because these types could also be used in a query without being
returned from them. With the previous implementation this would cause
DuckDB internal errors. With this change we do a final pass over the
optimized query plan to see if any expressions use such an unsupported
type, and error there too.

In passing this also removes an unintended `'` character from an error
message.